feat: extract link component
