+++ webcomponents/inline-editor/src/components/inline-editor/deckdeckgo-inline-editor.scss
@@ -39,31 +39,6 @@ div.deckgo-tools {
     height: var(--deckgo-inline-editor-sticky-height, 56px);
   }
 
-  div.link {
-    width: var(--deckgo-inline-editor-width, 216px);
-
-    height: 44px;
-    line-height: 44px;
-
-    z-index: var(--deckgo-inline-editor-link-zindex, 2);
-
-    input {
-      pointer-events: visible;
-
-      background: transparent;
-      width: 97%;
-      height: 100%;
-      color: var(--deckgo-inline-editor-link-color, #3880ff);
-      border: none;
-      outline: 0;
-      font-size: 16px;
-
-      &::placeholder {
-        color: var(--deckgo-inline-editor-link-placeholder-color, #3880ff);
-      }
-    }
-  }
-
   div.color {
     max-width: var(--deckgo-inline-editor-width, 216px);
