+++ webcomponents/inline-editor/src/components/link-actions/link-actions.scss
@@ -1,0 +1,24 @@
+:host {
+  width: var(--deckgo-inline-editor-width, 216px);
+
+  height: 44px;
+  line-height: 44px;
+
+  z-index: var(--deckgo-inline-editor-link-zindex, 2);
+
+  input {
+    pointer-events: visible;
+
+    background: transparent;
+    width: 97%;
+    height: 100%;
+    color: var(--deckgo-inline-editor-link-color, #3880ff);
+    border: none;
+    outline: 0;
+    font-size: 16px;
+
+    &::placeholder {
+      color: var(--deckgo-inline-editor-link-placeholder-color, #3880ff);
+    }
+  }
+}
