+++ webcomponents/inline-editor/src/index.html
@@ -1,48 +1,54 @@
 <!DOCTYPE html>
 <html dir="ltr" lang="en">
-<head>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0">
-  <title>DeckDeckGo - inline floating WYSIWYG HTML Editor Web Component</title>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0" />
+    <title>DeckDeckGo - inline floating WYSIWYG HTML Editor Web Component</title>
 
-  <script type="module" src="/build/deckdeckgo-inline-editor.esm.js"></script>
-  <script nomodule src="/build/deckdeckgo-inline-editor.js"></script>
+    <script type="module" src="/build/deckdeckgo-inline-editor.esm.js"></script>
+    <script nomodule src="/build/deckdeckgo-inline-editor.js"></script>
+  </head>
+  <body>
+    <p contenteditable>Bet on the Web</p>
 
-</head>
-<body>
+    <p contenteditable>Environment variables with StencilJS</p>
 
-  <p contenteditable>Bet on the Web</p>
+    <div>
+      <p contenteditable>
+        I noticed that the question regarding how to handle environment variables in Stencil’s projects or projects created with the Ionic PWA toolkit often
+        pops up 🤔
+      </p>
+    </div>
 
-  <p contenteditable>Environment variables with StencilJS</p>
+    <p contenteditable>
+      As I have implemented a solution to handle such parameters in the remote control of my project DeckDeckGo, the Progressive Web App alternative for simple
+      presentations, I thought about sharing my small implementation in this new article.
+    </p>
 
-  <div><p contenteditable>I noticed that the question regarding how to handle environment variables in Stencil’s projects or projects created with the Ionic PWA toolkit often pops up 🤔</p></div>
+    <h3 contenteditable>Credits</h3>
 
-  <p contenteditable>As I have implemented a solution to handle such parameters in the remote control of my project DeckDeckGo, the Progressive Web App alternative for simple presentations, I thought about sharing my small implementation in this new article.</p>
+    <p contenteditable>
+      The following solution was inspired by the one developed in the Ionic core project. One of the entry point for me was discovering the method setupConfig
+      in their source code. Therefore kudos to the awesome <a href="https://ionicframework.com">Ionic</a> team ❤️
+    </p>
 
-  <h3 contenteditable>Credits</h3>
+    <p contenteditable><img style="width: 100px;" src="https://deckdeckgo.com/assets/favicon/android-chrome-512x512.png" /></p>
 
-  <p contenteditable>The following solution was inspired by the one developed in the Ionic core project. One of the entry point for me was discovering the method setupConfig in their source code. Therefore kudos to the awesome <a href="https://ionicframework.com">Ionic</a> team ❤️</p>
+    <deckgo-inline-editor sticky-mobile="true" containers="h1,h2,h3,h4,h5,h6,p" img-editable="true" custom-actions="my-inline-action">
+      <span style="color: red;" slot="my-inline-action">My action</span>
+    </deckgo-inline-editor>
 
-  <p contenteditable><img style="width: 100px;" src="https://deckdeckgo.com/assets/favicon/android-chrome-512x512.png"/></p>
-
-  <deckgo-inline-editor sticky-mobile="true" containers="h1,h2,h3,h4,h5,h6,p" img-editable="true"
-                        custom-actions="my-inline-action">
-
-    <span style="color: red;" slot="my-inline-action">My action</span>
-
-  </deckgo-inline-editor>
-
-  <script type="text/javascript">
+    <script type="text/javascript">
       function myActions(e) {
-          if (e && e.detail && e.detail.action === 'my-inline-action') {
-              document.execCommand('insertHTML', false, `<acme-placeholder class="own-class">${e.detail.selection}</acme-placeholder>`);
-          }
+        if (e && e.detail && e.detail.action === 'my-inline-action') {
+          document.execCommand('insertHTML', false, `<acme-placeholder class="own-class">${e.detail.selection}</acme-placeholder>`);
+        }
       }
 
       document.addEventListener('customAction', myActions);
-  </script>
 
-
-
-</body>
+      document.addEventListener('imgDidChange', ($event) => console.log('imgDidChange', $event));
+      document.addEventListener('linkCreated', ($event) => console.log('linkCreated', $event));
+    </script>
+  </body>
 </html>
