+++ webcomponents/inline-editor/src/index.ts
@@ -1,2 +1,2 @@
 export * from './components';
-export * from './interfaces/deckdeckgo-inline-editor.interface';
+export * from './interfaces/interfaces';
