+++ webcomponents/inline-editor/src/interfaces/interfaces.ts
